move and rename stock page
